requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-PyMongo
python-dotenv
Flask-Dance
requests
